requirements.txt
asgiref==3.8.1
Django==5.1.1
Faker==30.6.0
Markdown==3.7
pillow==10.4.0
psycopg==3.2.2
psycopg-binary==3.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
